Refactor packages, fix enum deserialization
