+++ src/test/java/uk/gov/cabinetoffice/csl/integration/CancelEventTest.java
@@ -7,12 +7,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
+import org.springframework.http.MediaType;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.reactive.server.WebTestClient;
-import reactor.core.publisher.Mono;
+import org.springframework.web.reactive.function.BodyInserters;
 import uk.gov.cabinetoffice.csl.configuration.TestConfig;
 import uk.gov.cabinetoffice.csl.controller.model.CancelBookingDto;
 import uk.gov.cabinetoffice.csl.domain.learnerrecord.*;
+import uk.gov.cabinetoffice.csl.domain.learnerrecord.booking.BookingCancellationReason;
 import uk.gov.cabinetoffice.csl.domain.learningcatalogue.Course;
 import uk.gov.cabinetoffice.csl.domain.learningcatalogue.ModuleType;
 import uk.gov.cabinetoffice.csl.util.CSLServiceWireMockServer;
@@ -20,7 +22,6 @@ import uk.gov.cabinetoffice.csl.util.TestDataService;
 import uk.gov.cabinetoffice.csl.util.stub.CSLStubService;
 
 import java.math.BigDecimal;
-import java.net.URI;
 import java.util.List;
 
 @Slf4j
@@ -66,12 +67,6 @@ public class CancelEventTest extends CSLServiceWireMockServer {
         course.getModule(moduleId).setCost(BigDecimal.valueOf(0L));
         course.getModule(moduleId).setModuleType(ModuleType.facetoface);
         courseRecord.setState(State.REGISTERED);
-        BookingDto dto = BookingDto.builder()
-                .event(URI.create(String.format("http://localhost:9000/learning_catalogue/courses/%s/modules/%s/events/%s", courseId, moduleId, eventId)))
-                .status(BookingStatus.CANCELLED)
-                .learner(userId)
-                .learnerEmail(userEmail)
-                .learnerName("testName").build();
         List<PatchOp> expectedCourseRecordPatches = List.of(PatchOp.replacePatch("state", "UNREGISTERED"));
         List<PatchOp> expectedModuleRecordPatches = List.of(
                 PatchOp.replacePatch("state", "UNREGISTERED"),
@@ -83,16 +78,25 @@ public class CancelEventTest extends CSLServiceWireMockServer {
         String expectedCancellationJsonInput = """
                 {"cancellationReason": "PRIORITIES", "status":"Cancelled"}
                 """;
-        cslStubService.getLearnerRecord().cancelBooking(eventId, userId, expectedCancellationJsonInput, dto);
+        String event = String.format("http://localhost:9000/learning_catalogue/courses/%s/modules/%s/events/%s", courseId, moduleId, eventId);
+        String bookingDtoJsonResponse = String.format("""
+                {"event": "%s", "status":"Cancelled", "learner": "%s",
+                "learnerEmail": "%s", "learnerName":"testName", "cancellationReason": "Other work priorities"}
+                """, event, userId, userEmail);
+        cslStubService.getLearnerRecord().cancelBooking(eventId, userId, expectedCancellationJsonInput, bookingDtoJsonResponse);
         cslStubService.stubUpdateCourseRecord(courseId, course, userId, courseRecords,
                 1, expectedModuleRecordPatches, moduleRecord, expectedCourseRecordPatches, courseRecord);
         CancelBookingDto inputDto = new CancelBookingDto(BookingCancellationReason.PRIORITIES);
+        String inputJson = """
+                {"reason": "PRIORITIES"}
+                """;
         String url = String.format("/courses/%s/modules/%s/events/%s/cancel_booking", courseId, moduleId, eventId);
         webTestClient
                 .post()
                 .uri(url)
                 .header("Authorization", "Bearer fakeToken")
-                .body(Mono.just(inputDto), CancelBookingDto.class)
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(BodyInserters.fromValue(inputJson))
                 .exchange()
                 .expectStatus()
                 .is2xxSuccessful()
